requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
--trusted-host pypi.python.org
--trusted-host pypi.org
--trusted-host files.pythonhosted.org

absl-py==0.9.0
genicam==1.0.0            # via harvesters
harvesters==1.0.1
numpy==1.18.1
opencv-python==4.1.2.30
six==1.13.0               # via absl-py
